+++ CHANGELOG.md
@@ -8,10 +8,11 @@
 - [ ] Add `travis` to Gemfile, and run it from our local shims vs from global during `lanyon encrypt`
 - [ ] Add image optimizer from `assets/images` -> `assets/build/images`
 
-## v0.0.44 (Unreleased)
+## v0.0.44
 
 [Diff](https://github.com/kvz/lanyon/compare/v0.0.43...v0.0.44)
 
+- [x] Fix babel source maps
 - [x] Only exit on error when in production mode
 
 ## v0.0.43